+++ src/App/tests/test_getZopeVersion.py
@@ -15,9 +15,8 @@
 
 import unittest
 
-from pkg_resources import get_distribution
-
 from App.version_txt import getZopeVersion
+from pkg_resources import get_distribution
 
 
 class Test(unittest.TestCase):