+++ src/Testing/ZopeTestCase/ZopeLite.py
@@ -38,6 +38,9 @@ import Products  # NOQA
 import ZODB  # NOQA
 import Zope2  # NOQA
 import Zope2.Startup.run  # NOQA
+# Allow test authors to install Zope products into the test environment. Note
+# that installProduct() must be called at module level -- never from tests.
+from OFS.Application import get_folder_permissions  # NOQA; NOQA
 from OFS.Application import get_products
 from OFS.Application import install_package
 from OFS.Application import install_product
@@ -45,11 +48,6 @@ from OFS.Folder import Folder  # NOQA
 from Testing.ZopeTestCase import layer
 # ZODB sandbox factory
 from ZODB.DemoStorage import DemoStorage  # NOQA
-
-
-# Allow test authors to install Zope products into the test environment. Note
-# that installProduct() must be called at module level -- never from tests.
-from OFS.Application import get_folder_permissions  # NOQA; NOQA
 
 
 # Allow code to tell it is run by the test framework