+++ src/Testing/ZopeTestCase/testZODBCompat.py
@@ -308,7 +308,8 @@ class TestTransactionAbort(ZopeTestCase.ZopeTestCase):
 
 
 def test_suite():
-    from unittest import TestSuite, makeSuite
+    from unittest import TestSuite
+    from unittest import makeSuite
     suite = TestSuite()
     suite.addTest(makeSuite(TestCopyPaste))
     suite.addTest(makeSuite(TestImportExport))